Remove plurality derivation and discussion
The derivation is for the outdated (N,C) communicability-independent
game-type classification. While it is still correct
if the A/B game types are replaced with N/C communicabilities,
it then only makes predictions regarding the percentage of mixed games
compared to (neutral/cooperative) unmixed games. However, this is not
helpful for the well-mixed case, because it is virtually never
unsynchronized (i.e. order parameter is never in the 1/2 \pm 1/2\sqrt{N}
range). And the ~0.3% of time when it is unsynchronized,
we do not see the expected ~15% (*0.3%) mixed game percentage. On the
other hand, the derivation is not applicable to c-elegans;
the fact that we do not see ~15% mixed games in figure 6b (game type
figure for all games, *not* fig 6d for only mixed games)
means our assumption that c-elegans is a \overline{d}-random graph must
be false.

Additionally, it is easily extended to the new,
communicability-dependent case by replacing the complete graph
with a bipartite (N and C) graph. Then, the upper and lower limits
for plurality non-synchronous games becomes the point
1) when C switches from mostly A to mostly B, and 2) when N switches
from mostly A to mostly B. However, this appears to be unhelpful
for the well-mixed setup: the evolutionary dynamics means that
if the resident population is (WLOG) C with phase A and the invading
population is N with phase B, the low mutation rate
means C will always have phase A and B will always have phase B,
so the bipartite graph will *only* have edges with \Delta \theta = B - A
\neq 0. Indeed, figure 6c inset shows that the mixed game types
are distributed approximately equal to the fraction of white dots
in each region of phase space in figure 2b. It does seem there is a
slight bias towards the nearly-synchronous game type (esp. for alpha =
0.5, 0.75). However, it is unclear why this would be the case. Likewise,
this analysis is inapplicable to the c-elegans case: a brief glance
at any of the topology plots (e.g.fig 4d) shows that the assumption
of only two phases is entirely broken.